requirements.txt
appdirs==1.4.3
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
black==19.3b0
boto3==1.9.158
botocore==1.12.158
certifi==2018.11.29
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cryptography==3.4.7
defusedxml==0.5.0
dj-database-url==0.5.0
dj-rest-auth==2.1.7
Django==2.1.10
django-allauth==0.44.0
django-cors-headers==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.7.1
django-taggit==0.23.0
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
docutils==0.14
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.9.4
lazy-object-proxy==1.3.1
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==8.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.19
PyJWT==1.7.0
pylint==2.3.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.23.0
toml==0.10.0
typed-ast==1.4.3
typing_extensions
uritemplate==3.0.1
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.0
razorpay
